+++ include/EventListener.hh
@@ -10,8 +10,6 @@
 class EventListener {
  public:
   virtual ~EventListener() = default;
-  virtual void mouseDragStart() = 0;
-  virtual void mouseDragEnd() = 0;
   virtual void mouseDrag(int xrel, int yrel) = 0;
   virtual void mouseMotion(int x, int y) = 0;
   virtual void mouseWheel(int x, int y) = 0;