+++ include/EventLoop.hh
@@ -14,6 +14,7 @@ public:
   virtual void scheduleRefreshDisplay(int delay) = 0;
   virtual void scheduleQueueAudio(int delay) = 0;
   virtual void scheduleFade(int delay) = 0;
+  virtual void clearAdvanceFrame() = 0;
 };
 
 class EventLoop: public EventScheduler {