+++ src/main.cc
@@ -6,12 +6,12 @@
 #include <fstream>
 #include <iostream>
 #include <vector>
-#include "spdlog/cfg/env.h"
+
 #include "spdlog/spdlog.h"
-#include "spdlog/fmt/ostr.h"
+
 
 #include "VivictPP.hh"
-#include "ui/VivictUI.hh"
+#include "Controller.hh"
 #include "SourceConfig.hh"
 #include "vmaf/VmafLog.hh"
 
@@ -74,8 +74,7 @@ int main(int argc, char **argv) {
     std::vector<std::string> filters = {leftFilter, rightFilter};
     std::vector<std::string> vmafLogfiles = {leftVmaf, rightVmaf};
 
-    spdlog::cfg::load_env_levels();
-    spdlog::set_pattern("%H:%M:%S.%e %^%=8l%$ %-20n thread-%t  %v");
+    vivictpp::logging::initializeLogging();
 
     std::vector<SourceConfig> sourceConfigs;
     for (size_t i = 0; i<sources.size(); i++) {
@@ -89,8 +88,9 @@ int main(int argc, char **argv) {
     }
 
     VivictPPConfig vivictPPConfig(sourceConfigs, disableAudio);
-    vivictpp::ui::VivictUI vivictUI(vivictPPConfig);
-    return vivictUI.run();
+    auto sdlEventLoop = std::make_shared<vivictpp::sdl::SDLEventLoop>(vivictPPConfig.sourceConfigs);
+    vivictpp::Controller controller(sdlEventLoop, sdlEventLoop, vivictPPConfig);
+    return controller.run();
   } catch (const std::exception &e) {
     std::cerr << "Vivict had an unexpected error: " << e.what() << std::endl;
     return 1;