+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
+## Current
+
+### Fixed
+- Seeking backwards now stops on stream start
+- Playback stops on stream end
+- Seeking forward stops on stream end
+- Fixed problems with playback after multiple seeks
 
 ## x.y.z - YYYY-MM-DD